ignore non-exists bundles

+++ src/tauri_build_linux.ts
@@ -1,6 +1,7 @@
 import { TauriBuild } from './tauri_build'
 import * as core from '@actions/core'
 import { execPromise } from './utils'
+import { existsSync } from 'fs'
 
 class TauriBuildLinux extends TauriBuild {
   os_label: string = 'linux'
@@ -109,10 +110,6 @@ EOF`)
         to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}_amd64`
       },
       {
-        from: `src-tauri/target/x86_64-unknown-linux-gnu/release/bundle/appimage/${this.projectName}_${this.version}_amd64.AppImage`,
-        to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}_amd64.AppImage`
-      },
-      {
         from: `src-tauri/target/x86_64-unknown-linux-gnu/release/bundle/deb/${this.projectName}_${this.version}_amd64.deb`,
         to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}_amd64.deb`
       },
@@ -120,14 +117,14 @@ EOF`)
         from: `src-tauri/target/x86_64-unknown-linux-gnu/release/bundle/rpm/${this.projectName}-${this.version}-1.x86_64.rpm`,
         to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}.x86_64.rpm`
       },
+      {
+        from: `src-tauri/target/x86_64-unknown-linux-gnu/release/bundle/appimage/${this.projectName}_${this.version}_amd64.AppImage`,
+        to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}_amd64.AppImage`
+      },
       // i686
       {
         from: `src-tauri/target/i686-unknown-linux-gnu/release/${this.projectName}`,
         to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}_i386`
-      },
-      {
-        from: `src-tauri/target/i686-unknown-linux-gnu/release/bundle/appimage/${this.projectName}_${this.version}_i386.AppImage`,
-        to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}_i386.AppImage`
       },
       {
         from: `src-tauri/target/i686-unknown-linux-gnu/release/bundle/deb/${this.projectName}_${this.version}_i386.deb`,
@@ -136,6 +133,10 @@ EOF`)
       {
         from: `src-tauri/target/i686-unknown-linux-gnu/release/bundle/rpm/${this.projectName}-${this.version}-1.i386.rpm`,
         to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}.i386.rpm`
+      },
+      {
+        from: `src-tauri/target/i686-unknown-linux-gnu/release/bundle/appimage/${this.projectName}_${this.version}_i386.AppImage`,
+        to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}_i386.AppImage`
       },
       // aarch64
       {
@@ -150,6 +151,10 @@ EOF`)
         from: `src-tauri/target/aarch64-unknown-linux-gnu/release/bundle/rpm/${this.projectName}-${this.version}-1.aarch64.rpm`,
         to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}.aarch64.rpm`
       },
+      {
+        from: `src-tauri/target/aarch64-unknown-linux-gnu/release/bundle/appimage/${this.projectName}_${this.version}_aarch64.AppImage`,
+        to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}_aarch64.AppImage`
+      },
       // armv7
       {
         from: `src-tauri/target/armv7-unknown-linux-gnueabihf/release/${this.projectName}`,
@@ -162,10 +167,20 @@ EOF`)
       {
         from: `src-tauri/target/armv7-unknown-linux-gnueabihf/release/bundle/rpm/${this.projectName}-${this.version}-1.armhfp.rpm`,
         to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}.armhfp.rpm`
+      },
+      {
+        from: `src-tauri/target/armv7-unknown-linux-gnueabihf/release/bundle/appimage/${this.projectName}_${this.version}_armhfp.AppImage`,
+        to: `${this.outputDir}/${this.projectName}-${this.os_label}-${this.version}_armhfp.AppImage`
       }
     ]
     for (const file of filesToCopy) {
-      await execPromise(`cp ${file.from} ${file.to}`)
+      // check if the source file exists
+      if (existsSync(file.from)) {
+        core.info(`Copying ${file.from} to ${file.to}`)
+        await execPromise(`cp ${file.from} ${file.to}`)
+      } else {
+        core.warning(`Source file ${file.from} does not exist`)
+      }
     }
     // Build for android after cleanup target directory
     await execPromise('rm -rf src-tauri/target')
